made component from a singular event
